Better packaging
Moved everything into a \bqt folder and renamed bqt.py to __init__.py. Updated the imports on the blender_applications\*.py files to match. Moved License and Readme.md into bqt\ along with the stylesheet and icon.

Now all a user has to do is copy bqt\ to their blender scripts\startup folder and they are good to go.