Attempt to reduce Alt-Tabbing bug in bqt
Attempting to fix the Alt-Hang when Alt-Tabbing between Blender and other applications by running a modal operator that checks if the window just got focus back. If that is the case, check if Blender perceives an event containing CTRL/ALT/SHIFT and fire a release event on the keyboard. Does not catch all cases.

# Conflicts:
#	bqt/blender_applications/win32_blender_application.py
#	requirements.txt